pyproject.toml
[project]
name = "try"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "joblib",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.2",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
    "torchaudio",
    "torchvision>=0.23.0",
    "ttkbootstrap>=1.12.1",
]
[tool.uv]
environments = [ "sys_platform == 'win32'"]
default-index = "https://pypi.org/simple"

extra-index-url = ["https://download.pytorch.org/whl/cu126"]
[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
default = true

